refactor(interp): unify cfg and trace jit lowering
